requirements.txt
numpy==1.24.3
onnxruntime==1.23.2
Pillow==10.4.0
scikit-learn==1.6.0
joblib==1.4.2
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
python-multipart==0.0.20
streamlit==1.41.1
pandas==2.2.3
sqlalchemy==2.0.36
openai==1.59.3
ollama==0.4.7
python-dotenv==1.0.1
tensorflow>=2.16
tensorboard>=2.16